+++ site/FrontPage.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { Head } from "./Head"
 import { SiteHeader } from "./SiteHeader"
 import { SiteFooter } from "./SiteFooter"
-import { CategoryWithEntries, FullPost, EntryNode } from "db/wpdb"
+import { FullPost, EntryNode } from "db/wpdb"
 import { faRss } from "@fortawesome/free-solid-svg-icons/faRss"
 import { faAngleRight } from "@fortawesome/free-solid-svg-icons/faAngleRight"
 import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons/faExternalLinkAlt"
@@ -15,6 +15,7 @@ import {
     NewsletterSubscriptionForm,
     NewsletterSubscriptionContext,
 } from "site/NewsletterSubscription"
+import { CategoryWithEntries } from "clientUtils/owidTypes"
 
 const splitOnLastWord = (str: string) => {
     const endIndex = (str.lastIndexOf(" ") as number) + 1
